pyproject.toml
[tool.poetry]
name = "turlington"
version = "0.1.0"
description = ""
authors = ["uncleiansreadyrice <128861127+uncleiansreadyrice@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.2"
ampapi = "^1.0.8"
mcstatus = "^10.0.2"
jishaku = "^2.5.1"
requests = "^2.28.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
